refactored into _db.py to remove circular import dependencies
