+++ src/server/_common.py
@@ -2,23 +2,17 @@ from typing import cast
 import time
 
 from flask import Flask, g, request
-from sqlalchemy import create_engine, event, MetaData
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
+from sqlalchemy import event
 from sqlalchemy.engine import Connection, Engine
 from werkzeug.local import LocalProxy
 
 from delphi.epidata.common.logger import get_structured_logger
-from ._config import SECRET, SQLALCHEMY_DATABASE_URI, SQLALCHEMY_ENGINE_OPTIONS, REVERSE_PROXIED
+from ._config import SECRET, REVERSE_PROXIED
+from ._db import engine
 from ._exceptions import DatabaseErrorException, EpiDataException
 from ._security import _get_current_user, _is_public_route, resolve_auth_token, update_key_last_time_used
 
-engine: Engine = create_engine(SQLALCHEMY_DATABASE_URI, **SQLALCHEMY_ENGINE_OPTIONS)
-metadata = MetaData(bind=engine)
-Base = declarative_base()
 
-Session = sessionmaker(bind=engine)
-session = Session()
 
 app = Flask("EpiData", static_url_path="")
 # for example if the request goes through one proxy
